Specify directory as data dir.
Set's the data dir as residing at /data.  The Dockerfile also sets a
VOLUME["/data"] to facilitate the assigning of a real directory.
Eg., "-v /var/docker/gitbucket:/data:rw"

The '-v' means we are assigning a volume.  It allows us to make host
directories available to processes inside the container.

Using the config at the beginning, we will be mapping the local
directory (`pwd`/data) to a container's local directory (/data), and
making the data read/write (rw).